Bump to JDK 17 (newest LTS)
Migrate JSON-IO to Jackson
Remove path limitation on Windows

+++ src/core/master/src/main/java/org/drftpd/master/commands/dataconnection/DataConnectionHandler.java
@@ -784,7 +784,7 @@ public class DataConnectionHandler extends CommandInterface {
 
             // Checks maxsim up/down
             // _simup OR _simdown = 0, exempt
-            int comparison;
+            double comparison;
             int count = BaseFtpConnection.countTransfersForUser(conn.getUserNull(), direction);
             env.put("maxsim", count);
